prompt_test_version-main-2/requirements.txt
openai
boto3
fastapi
uvicorn[standard]
google-cloud-texttospeech